+++ src/globals/Header/components/accordion.tsx
@@ -5,7 +5,7 @@ import { ChevronDownIcon } from 'lucide-react';
 import * as React from 'react';
 
 import { IconNames } from '@/fields/lucideIconPicker/types';
-import { buttonVariants } from '@/primitives/ui/button-prime';
+import { buttonVariants } from '@/primitives/ui/button';
 import { cn } from '@/utilities/cn';
 import { cva, VariantProps } from 'class-variance-authority';
 import DynamicIcon from '../../../components/DynamicIcon';
